requirements.txt
click==8.0.1
DateTime==4.3
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
zope.interface==5.4.0
